Implement arguments parsing

+++ ap03/include/fatorial.h
@@ -1,8 +1,8 @@
 #ifndef FATORIAL_H
 #define FATORIAL_H
 
-int fatorial_recursivo(int n);
-int fatorial_iterativo(int n);
+long long fatorial_recursivo(int n);
+long long fatorial_iterativo(int n);
 
 
 #endif /* FATORIAL_H */
