+++ ap03/include/fibo.h
@@ -1,7 +1,7 @@
 #ifndef FIBO_H
 #define FIBO_h
 
-int fibo_recursivo(int n);
-int fibo_iterativo(int n);
+long long fibo_recursivo(int n);
+long long fibo_iterativo(int n);
 
 #endif /* FIBO_H */
